MOVE hooks into lib/addons
