+++ lib/addons/package.json
@@ -23,8 +23,8 @@
   "dependencies": {
     "@storybook/api": "5.2.0-beta.40",
     "@storybook/channels": "5.2.0-beta.40",
-    "@storybook/client-api": "5.2.0-beta.40",
     "@storybook/client-logger": "5.2.0-beta.40",
+    "@storybook/core-events": "5.2.0-beta.40",
     "core-js": "^3.0.1",
     "global": "^4.3.2",
     "util-deprecate": "^1.0.2"
