+++ lib/addons/src/public_api.ts
@@ -9,4 +9,6 @@ export * from './make-decorator';
 export * from './index';
 export * from './types';
 export * from './storybook-channel-mock';
+export * from './hooks';
+
 export default addons;
