+++ lib/client-api/src/hooks.test.js
@@ -1,4 +1,5 @@
 import { FORCE_RE_RENDER, STORY_RENDERED } from '@storybook/core-events';
+import addons from '@storybook/addons';
 import { defaultDecorateStory } from './client_api';
 import {
   applyHooks,
@@ -32,11 +33,8 @@ beforeEach(() => {
     removeListener: jest.fn(),
   };
   hooks = new HooksContext();
+  addons.setChannel(mockChannel);
 });
-
-jest.mock('@storybook/addons', () => ({
-  getChannel: () => mockChannel,
-}));
 
 const decorateStory = applyHooks(defaultDecorateStory);
